+++ sdk/keyvault/azure-security-keyvault-secrets/CMakeLists.txt
@@ -45,10 +45,10 @@ set(
     src/private/keyvault_protocol.hpp
     src/private/secret_constants.hpp
     src/private/secret_serializers.hpp
-    inc/azure/keyvault/secrets/keyvault_operations.hpp
     src/keyvault_protocol.cpp
     src/secret_client.cpp
     src/secret_serializers.cpp
+    src/keyvault_operations.cpp
 )
 
 add_library(azure-security-keyvault-secrets ${AZURE_SECURITY_KEYVAULT_SECRETS_HEADER} ${AZURE_SECURITY_KEYVAULT_SECRETS_SOURCE})
