+++ sdk/keyvault/azure-security-keyvault-secrets/inc/azure/keyvault/secrets/secret_client.hpp
@@ -10,23 +10,19 @@
 
 #include "azure/keyvault/secrets/keyvault_backup_secret.hpp"
 #include "azure/keyvault/secrets/keyvault_deleted_secret.hpp"
+#include "azure/keyvault/secrets/keyvault_operations.hpp"
 #include "azure/keyvault/secrets/keyvault_options.hpp"
 #include "azure/keyvault/secrets/keyvault_secret.hpp"
 #include "dll_import_export.hpp"
 #include <azure/core/http/http.hpp>
 #include <azure/core/internal/http/pipeline.hpp>
 #include <azure/core/response.hpp>
-
 #include <stdint.h>
 #include <string>
 
 namespace Azure { namespace Security { namespace KeyVault { namespace _detail {
   class KeyVaultProtocolClient;
 }}}} // namespace Azure::Security::KeyVault::_detail
-namespace Azure { namespace Security { namespace KeyVault { namespace Secrets {
-  class KeyVaultRestoreDeletedSecretOperation;
-  class KeyVaultDeleteSecretOperation;
-}}}} // namespace Azure::Security::KeyVault::Secrets
 
 namespace Azure { namespace Security { namespace KeyVault { namespace Secrets {
